processing/requirements.txt
pyyaml==5.3.1
awscli==1.19.53
boto3==1.17.53
bs4==0.0.1
eoepca-scim==2.8.1
folium==0.12.1
geolinks==0.2.0
ipyleaflet==0.17.1
jsonschema==3.2.0
lxml==4.6.3
matplotlib==3.7.1
OWSLib @ git+https://github.com/geopython/OWSLib@792fa034dc3f44af8fea62477d71466ca3d3a06b
pycrypto==2.6.1
pyjwkest==1.4.2
pystac-client==0.5.1
rauth==0.7.3
robotframework==3.2.2
tifffile==2021.4.8
WellKnownHandler==0.2.0
